AMVP-1471 Add GPIO Reset Line For RX Packet FIFO In MAC Core
Add ability to reset FIFO between MAC and MCDMA channel. If this
FIFO overruns, it needs to be reset to bring it into service again.
The FIFO must come up out of reset to allow clocks to stablize. Once
clocks are stable the FIFO is placed in reset until the link is
brought into service.