AMVP-1481 PTP TX Timestamps Get Out Of Sync
The TX DMA done interrupt occurs ahead of the packet being actually
transmitted by the hardware, therefore the TX TS is placed in the
FIFO sometime after the DMA is complete. Usually this delay is
short enough that by the time the interrupt routine runs, the TX TS
FIFO is populated. But occasionaly, it is not. This patch fixes this
and the way the FIFO is accessed to conform to the PG080 specification:

1. Queue all completed PTP messages before processing TX TS timestamp.
2. Enhance error recovery to cope with not only the timestamp
   FIFO retarding (head of FIFO does not match current packet) but
   also dealing with lost timestamp.
3. Ensure FIFO occupancy (RFO) is checked before packet length (RLR)
   register is read. If RLR is read and empty, then an underrun
   error interrupt status is asserted. In this condition, the contents
   of the RDFD is not guaranteed to be correct.
4. Turned on sequential tags per port to aid in debugging of lost
   timestamp interrupts.

The hardware is still failing to provide TX timestamps every 3-30
seconds across 12 radio ports. This will be the focus of the next
round of troubleshooting. This patch allows the PTP timestamping to
recovery after a temporary loss of sync amounting to 15us.