Remove TSN Driver From Xilinx Driver Directory (same as upstream v6.1 kernel)
The TSN driver was moved from drivers/net/ethernet/xilinx directory to
drivers/staging/xilinx-tsn in upstream v6.1 Xilinx driver. To ease the
backporting of v6.x features to v5.10 kernel, remove the TSN driver in
the Antevia v5.10 zu19 branch. We do not use it, and will not in the future.